+++ cli/src/commands/util/gc.rs
@@ -27,9 +27,6 @@ use crate::ui::Ui;
 ///
 /// To garbage-collect old operations and the commits/objects referenced by
 /// them, run `jj op abandon ..<some old operation>` before `jj util gc`.
-///
-/// Previous versions of a change that are reachable via the evolution log are
-/// not garbage-collected.
 #[derive(clap::Args, Clone, Debug)]
 pub struct UtilGcArgs {
     /// Time threshold